requirements.txt
pygame==2.6.1
PyAudio==0.2.13
SpeechRecognition==3.9.0
pyttsx3==2.90
gTTS==2.4.3
requests==2.31.0
google-generativeai==0.23.0
musiclibrary==0.1.0
